+++ include/vec/magic.h
@@ -1,0 +1,8 @@
+#ifndef VEC_MAGIC_H_IS_INCLUDED
+#define VEC_MAGIC_H_IS_INCLUDED
+
+#include "vec/types.h"
+
+
+
+#endif