+++ include/vec/defaults.h
@@ -1,0 +1,8 @@
+#ifndef VEC_DEFAULTS_H_IS_INCLUDED
+#define VEC_DEFAULTS_H_IS_INCLUDED
+
+#include "vec/types.h"
+
+
+
+#endif