+++ lib/defaults.c
@@ -1,0 +1,5 @@
+#include "vec/defaults.h"
+
+
+
+