+++ include/vec/types.h
@@ -64,6 +64,7 @@ struct VecNamespace {
 	VecNamespace (*erase) (Vec v, long start, long end);
 
 	VecNamespace (*print) (Vec v);
+	VecNamespace (*print_slots) (Vec v);
 	VecNamespace (*print_stats) (Vec v);
 
 	VecNamespace (*swap) (Vec v, long a, long b);