+++ lib/vec.c
@@ -41,6 +41,7 @@ struct VecNamespace vec_struct = {
 	.full = &vec_full,
 
 	.print = &vec_print,
+	.print_slots = &vec_print_slots,
 	.print_stats = &vec_print_stats,
 
 	.swap = &vec_swap,